requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.0
anthropic>=0.25.0
openai>=1.0.0
gradio>=4.0.0
